Move `compute_output_shape` to new `ComputeOutputShape` trait.
