+++ apollo-federation/src/sources/connect/json_selection/methods/tests.rs
@@ -1,3 +1,4 @@
+use apollo_compiler::collections::IndexMap;
 use insta::assert_debug_snapshot;
 use serde_json_bytes::json;
 
@@ -211,10 +212,9 @@ fn test_map_method() {
             json_selection.apply_to(&single_value_data),
             (Some(json!(["123"])), vec![]),
         );
-        let output_shape = json_selection.compute_output_shape(
-            Shape::from_json_bytes(&single_value_data),
-            &IndexMap::default(),
-        );
+        let mut named_shapes = IndexMap::default();
+        named_shapes.insert("$root".to_string(), Shape::from_json_bytes(&single_value_data));
+        let output_shape = json_selection.output_shape(&named_shapes);
         assert_eq!(output_shape.pretty_print(), "List<String>");
     }
 }
